LayoutTests/imported/w3c/web-platform-tests/content-security-policy/media-src/media-src-7_2_2.sub-expected.txt

Audio element src attribute must match src list - negative test


PASS Disallaowed audio src
PASS Disallowed audio source element
FAIL Test that securitypolicyviolation events are fired assert_equals: expected "http://www2.localhost:8800/media/sound_5.oga" but got "http://www2.localhost:8800"
